Fix: workaround postinstall issue
`npm` executes the postinstall script not only when the current module
is installed but also when its dependencies are installed. In the latter
case, the source tree had not been adequately prepared and the build
failed.

There is a package hook called "prepare" and even though it sounds like
what we want it actually isn't - it only gets triggered before a pack or
publish.